// src/platform/packages/shared/shared-ux/document_data_cascade/impl/src/components/data_cascade_impl/types.d.ts
export type CascadeSizing = keyof Pick<EuiThemeShape['size'], 's' | 'm' | 'l'>;
export interface CascadeGroupNodeUIInteraction<G extends GroupNode> {
    /**
     * The row instance that was interacted with in the group by hierarchy.
     */
    row: G;
    /**
     * The path of the row that was interacted with in the group by hierarchy.
     *
     * @example
     * ```
     * ['@timestamp', 'geo.src.country_name']
     * ```
     */
    nodePath: string[];
    /**
     * KV record of the path values for the row node.
     *
     * @example
     * ```
     * {
     *   '@timestamp': '2026-01-13T00:00:00.000Z',
     *   'geo.src.country_name': 'United States'
     * }
     * ```
     */
    nodePathMap: Record<string, string>;
}
type OnCascadeLeafNodeExpandedArgs<G extends GroupNode> = CascadeGroupNodeUIInteraction<G>;
type OnCascadeLeafNodeCollapsedArgs<G extends GroupNode> = CascadeGroupNodeUIInteraction<G>;
export interface CascadeRowCellRendererProps<G extends GroupNode, L extends LeafNode> {
    data: L[] | null;
    cellId: string;
    nodePath: string[];
    virtualizerController: ChildVirtualizerController;
    rowIndex: number;
}
export interface CascadeRowCellPrimitiveProps<G extends GroupNode, L extends LeafNode> extends CellContext<G, unknown> {
    /**
     * Size of the row cell
     */
    size: CascadeSizing;
    /**
     * Callback invoked when a leaf node gets expanded, which can be used to fetch data for leaf nodes.
     */
    onCascadeLeafNodeExpanded: (args: OnCascadeLeafNodeExpandedArgs<G>) => Promise<L[]>;
    /**
     * Callback invoked when a leaf node gets collapsed, possibly to clean up any data associated with the leaf node or cancel any pending requests if necessary.
     */
    onCascadeLeafNodeCollapsed?: (args: OnCascadeLeafNodeCollapsedArgs<G>) => void;
    getVirtualizer: () => CascadeRootVirtualizerReturnValue;
    /**
     * Render prop function that provides the leaf node data when available, which can be used to render the content we'd to display with the data received.
     */
    children: (args: CascadeRowCellRendererProps<G, L>) => React.ReactNode;
}